Move methods related to its own classes
