+++ src/plugins/views/editors/Flow/Components/interface/MainInterface.js
@@ -370,9 +370,7 @@ export default class MainInterface {
   onSelectNode = data => {
     const { nodes, shiftKey } = data;
     const { selectedNodes } = this;
-    const filterNodes = nodes.filter(
-      n => n.data.model !== StartNode.model
-    );
+    const filterNodes = nodes.filter(n => n.data.model !== StartNode.model);
 
     this.selectedLink = null;
 
@@ -410,6 +408,25 @@ export default class MainInterface {
         this.hideLinks(node, visitedLinks);
       }
     });
+  };
+
+  onResetZoom = () => {
+    this.canvas.onResetZoom();
+  };
+
+  onMoveNode = event => {
+    const currentZoom = this.canvas.currentZoom?.k ?? 1;
+    const step = 2 / currentZoom + 1;
+    const delta = {
+      ArrowRight: [1 * step, 0],
+      ArrowLeft: [-1 * step, 0],
+      ArrowUp: [0, -1 * step],
+      ArrowDown: [0, 1 * step]
+    };
+    const [dx, dy] = delta[event.code];
+    const [x, y] = [50, 50]; // skip boundaries validation used when dragging a node
+    this.graph.onNodeDrag(null, { x, y, dx, dy });
+    this.onDragEnd();
   };
 
   destroy = () => {
